roughly able to send email
